'cmap' format 4 not using correct counts for startCount, others.
Changed to use segCountX2/2 for the other 3 variables.  Also removed
the additional spaces from "segCountX2 / 2", as this caused strange
parse errors on following definition lines.

Also found that the OFFSET to _subtable's was not being computed correctly
and so changed the form for that line until working.  This affected all
cmap format subtables.