'cmap' extra array removed
Removed "_subtable[numTables] subtables" at end of the cmap table
definition as unnecessary.

This was reading the encoding tables again, apparently to "finish reading"
to the end of the cmap table.  However there is no guarantee that there
are 'numTables' count of tables there, only that there will be 'numTables'
count of 'encodingRecord' *entries*.

This is because the encodingRecord entries may re-use tables, as seen in
this cmap dump:
       0000                      version
       0005                      numTables
       0000   0003   0000 0224   platformID encodingID   offset
       0000   0004   0000 10ea
       0001   0000   0000 002c
       0003   0001   0000 0224
       0003   000a   0000 10ea
Two of the entries re-use tables, so while there are five entries,
there are only three encoding tables.