Implement 'GPOS' table and subtables
Some subtables common to GPOS/GSUB/JSTF were not separated out from
the GSUB table definition as initially implemented.  The definitions
found in the Layout Common Table Formats document are moved higher
and expanded.
   _ScriptList
   _FeatureList
   _ClassDefTable
   _CoverageTable

DefaultLangSys was not understood to be an offset to a subtable.  That
has been fixed, but that fix then required moving _LangSysTable higher
to be referenced by both DefaultLangSys and LangSys.

Added the common GPOS tables Anchor and MarkArray to beginning of GPOS
definition.  Unfortunately the ValueRecord definitions are excessively
wonderful and have to be repeated at each individual usage. (Referencing
the correct controlling ValueFormat flags value is difficult)

Added the GPOS Lookup subtables.
   Seen correct at least once:
        LookupType 2: Pair Adjustment Positioning Subtable
        LookupType 4: MarkToBase Attachment Positioning Subtable
        LookupType 6: MarkToMark Attachment Positioning Subtable
   Not yet seen:
        LookupType 1: Single Adjustment Positioning Subtable
        LookupType 3: Cursive Attachment Positioning Subtable
        LookupType 5: MarkToLigature Attachment Positioning Subtable
        LookupType 7: Contextual Positioning Subtables
        LookupType 8: Chaining Contextual Positioning Subtable
        LookupType 9: Extension Positioning

The implementation of GPOS LookupType 9 is a guess to be sure.

There is apparently no way to express "array of OFFSETs".  This caused
problems for
    GPOS LookupType 4: MarkToBase Attachment Positioning Subtable
    GPOS LookupType 5: MarkToLigature Attachment Positioning Subtable
    GPOS LookupType 6: MarkToMark Attachment Positioning Subtable
    GPOS LookupType 7: Contextual Positioning Subtables          Format 3
    GPOS LookupType 8: Chaining Contextual Positioning Subtable  Format 3
The 'solution' implemented was to make the array an array of entries, where
each entry would then say it was an OFFSET.  This introduces an additional
layer of object references.